server/requirements.txt
pandas
gspread
google-auth
ortools
python-dotenv
flask
flask-cors